drm/amdkfd: Improve the debug message
Some misleading debug messages are rectified. Moreover, the size
parameter used when allocating memory is changed to uint64_t from
size_t in order to be consistent with the size type in the alloc
ioctl parameter.

+++ drivers/gpu/drm/radeon/radeon_kfd.c
@@ -89,7 +89,7 @@ static int map_memory_to_gpu(struct kgd_dev *kgd, struct kgd_mem *mem,
 		void *vm);
 static int unmap_memory_from_gpu(struct kgd_dev *kgd, struct kgd_mem *mem,
 		void *vm);
-static int alloc_memory_of_gpu(struct kgd_dev *kgd, uint64_t va, size_t size,
+static int alloc_memory_of_gpu(struct kgd_dev *kgd, uint64_t va, uint64_t size,
 		void *vm, struct kgd_mem **mem,
 		uint64_t *offset, void **kptr,
 		struct kfd_process_device *pdd, uint32_t flags);
@@ -1395,7 +1395,7 @@ static int alloc_memory_of_scratch(struct kgd_dev *kgd,
 	return 0;
 }
 
-static int alloc_memory_of_gpu(struct kgd_dev *kgd, uint64_t va, size_t size,
+static int alloc_memory_of_gpu(struct kgd_dev *kgd, uint64_t va, uint64_t size,
 		void *vm, struct kgd_mem **mem,
 		uint64_t *offset, void **kptr,
 		struct kfd_process_device *pdd, uint32_t flags)
@@ -1427,7 +1427,7 @@ static int alloc_memory_of_gpu(struct kgd_dev *kgd, uint64_t va, size_t size,
 		goto err_bo_create;
 	}
 
-	pr_debug("Created BO on VRAM with size %zu bytes\n", size);
+	pr_debug("Created BO on VRAM with size %llu bytes\n", size);
 	bo->pdd = pdd;
 	ret = add_bo_to_vm(rdev, va, vm, bo, &bo_va);
 	if (ret != 0)
